+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - (Bugfix) Fix ArangoRoute Target switch in case of temporary error
 - (Bugfix) Fix IntOrString Schema Type
 - (Feature) Enable Operator Handler StackTraces
+- (Feature) Add spec validation for MLExtension
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD